Update makedep.py to support python 3
Currently, CVMIX can only be built (at least via MPAS-Ocean) with
python 2.7.  Given that python 2 is going away later this year,
it seems necessary to support python 3 as well.

With this merge, the python environment used to build CVMIx will
need the future package in python 2.7.  If this dependency is a
problem, it could be removed with the side effect that print
output will look strange in python 2.7 (be surrounded by
parentheses).